+++ src/wavefunctions/wavefunction.rs
@@ -1,7 +1,3 @@
 pub trait Wavefunction {
-    fn new(
-        num_particles: u32,
-        num_dimensions: u32,
-        num_parameters: u32,
-    ) -> Self; // Constructor for a Wavefunction
+    fn new(num_particles: u32, num_dimensions: u8, num_parameters: u8) -> Self; // Constructor for a Wavefunction
 }
